backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
gitpython==3.1.40
openai>=1.50.0
python-multipart==0.0.6
pydantic==2.5.0
